[refactor] ActivitySegment: detach ActivitySegment.from to mapper class with extension function
