+++ src/jvmMain/kotlin/uk/ryanwong/gmap2ics/app/models/timeline/Timeline.kt
@@ -4,28 +4,23 @@
 
 package uk.ryanwong.gmap2ics.app.models.timeline
 
-import uk.ryanwong.gmap2ics.app.models.timeline.activity.ActivitySegment
+import uk.ryanwong.gmap2ics.app.models.timeline.activity.toDomainModel
 import uk.ryanwong.gmap2ics.app.models.timeline.placevisit.PlaceVisit
 import uk.ryanwong.gmap2ics.app.utils.timezonemap.TimeZoneMapWrapper
 import uk.ryanwong.gmap2ics.data.models.timeline.TimelineObjects
 
 data class Timeline(
-    val timelineEntries: List<TimelineEntry>
+    val timelineEntries: List<TimelineEntry>,
 ) {
     companion object {
         fun from(
             timelineObjects: TimelineObjects,
-            timeZoneMap: TimeZoneMapWrapper
+            timeZoneMap: TimeZoneMapWrapper,
         ): Timeline {
             return Timeline(
                 timelineEntries = timelineObjects.timelineObjects?.map { timelineObject ->
                     TimelineEntry(
-                        activitySegment = timelineObject.activitySegment?.let { activitySegment ->
-                            ActivitySegment.from(
-                                activitySegmentDataModel = activitySegment,
-                                timeZoneMap = timeZoneMap
-                            )
-                        },
+                        activitySegment = timelineObject.activitySegment?.toDomainModel(timeZoneMap = timeZoneMap),
                         placeVisit = timelineObject.placeVisit?.let {
                             PlaceVisit.from(placeVisitDataModel = it, timeZoneMap = timeZoneMap)
                         }
